moving test utilities to make them available to hellbender-dataflow
