+++ src/main/java/org/broadinstitute/hellbender/utils/test/IntegrationTestSpec.java
@@ -1,11 +1,9 @@
-package org.broadinstitute.hellbender.tools;
+package org.broadinstitute.hellbender.utils.test;
 
 import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.hellbender.CommandLineProgramTest;
 import org.broadinstitute.hellbender.exceptions.GATKException;
 import org.broadinstitute.hellbender.utils.Utils;
-import org.broadinstitute.hellbender.utils.read.SamAssertionUtils;
-import org.broadinstitute.hellbender.utils.test.BaseTest;
 import org.broadinstitute.hellbender.utils.text.XReadLines;
 import org.testng.Assert;
