+++ src/main/java/org/broadinstitute/hellbender/utils/test/testers/SamFileTester.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.hellbender.utils.read.testers;
+package org.broadinstitute.hellbender.utils.test.testers;
 
 import htsjdk.samtools.*;
 import htsjdk.samtools.util.CloseableIterator;
