+++ src/test/java/org/broadinstitute/hellbender/tools/examples/ExampleIntervalWalkerIntegrationTest.java
@@ -1,7 +1,7 @@
 package org.broadinstitute.hellbender.tools.examples;
 
 import org.broadinstitute.hellbender.CommandLineProgramTest;
-import org.broadinstitute.hellbender.tools.IntegrationTestSpec;
+import org.broadinstitute.hellbender.utils.test.IntegrationTestSpec;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
