+++ src/test/java/org/broadinstitute/hellbender/tools/picard/analysis/CollectBaseDistributionByCycleIntegrationTest.java
@@ -2,7 +2,7 @@ package org.broadinstitute.hellbender.tools.picard.analysis;
 
 import htsjdk.samtools.metrics.MetricsFile;
     import org.broadinstitute.hellbender.CommandLineProgramTest;
-    import org.broadinstitute.hellbender.tools.IntegrationTestSpec;
+    import org.broadinstitute.hellbender.utils.test.IntegrationTestSpec;
     import org.testng.Assert;
     import org.testng.annotations.Test;
