+++ src/test/java/org/broadinstitute/hellbender/tools/picard/analysis/directed/CollectJumpingLibraryMetricsIntegrationTest.java
@@ -1,7 +1,7 @@
 package org.broadinstitute.hellbender.tools.picard.analysis.directed;
 
 import org.broadinstitute.hellbender.CommandLineProgramTest;
-import org.broadinstitute.hellbender.tools.IntegrationTestSpec;
+import org.broadinstitute.hellbender.utils.test.IntegrationTestSpec;
 import org.broadinstitute.hellbender.utils.test.BaseTest;
 import org.testng.annotations.Test;
