+++ src/test/java/org/broadinstitute/hellbender/tools/picard/sam/markduplicates/MarkDuplicatesIntegrationTest.java
@@ -5,8 +5,8 @@ import htsjdk.samtools.util.CloserUtil;
 import htsjdk.samtools.util.CollectionUtil;
 import htsjdk.samtools.util.IOUtil;
 import htsjdk.samtools.util.TestUtil;
-import org.broadinstitute.hellbender.utils.read.markduplicates.AbstractMarkDuplicatesCommandLineProgramTest;
-import org.broadinstitute.hellbender.utils.read.markduplicates.AbstractMarkDuplicatesTester;
+import org.broadinstitute.hellbender.utils.test.testers.AbstractMarkDuplicatesCommandLineProgramTest;
+import org.broadinstitute.hellbender.utils.test.testers.AbstractMarkDuplicatesTester;
 import org.broadinstitute.hellbender.utils.read.markduplicates.MarkDuplicatesTester;
 import org.broadinstitute.hellbender.utils.test.ArgumentsBuilder;
 import org.testng.Assert;
