+++ src/test/java/org/broadinstitute/hellbender/tools/spark/pipelines/FlagStatSparkIntegrationTest.java
@@ -3,7 +3,7 @@ package org.broadinstitute.hellbender.tools.spark.pipelines;
 import com.google.common.collect.Lists;
 import org.broadinstitute.hellbender.CommandLineProgramTest;
 import org.broadinstitute.hellbender.cmdline.StandardArgumentDefinitions;
-import org.broadinstitute.hellbender.tools.IntegrationTestSpec;
+import org.broadinstitute.hellbender.utils.test.IntegrationTestSpec;
 import org.broadinstitute.hellbender.utils.test.ArgumentsBuilder;
 import org.testng.Assert;
 import org.testng.annotations.Test;
