+++ src/test/java/org/broadinstitute/hellbender/utils/read/markduplicates/MarkDuplicatesSparkTester.java
@@ -2,6 +2,7 @@ package org.broadinstitute.hellbender.utils.read.markduplicates;
 
 import org.broadinstitute.hellbender.cmdline.CommandLineProgram;
 import org.broadinstitute.hellbender.tools.spark.transforms.markduplicates.MarkDuplicatesSpark;
+import org.broadinstitute.hellbender.utils.test.testers.AbstractMarkDuplicatesTester;
 
 /**
  * A tester class for {@link MarkDuplicatesSpark}.
